feat: Refactor code to separate interfaces and mock data
